LPS-65920 Move ChatPortletKeys to api module
