+++ modules/apps/chat/chat-api/bnd.bnd
@@ -2,6 +2,7 @@ Bundle-Name: Liferay Chat API
 Bundle-SymbolicName: com.liferay.chat.api
 Bundle-Version: 1.0.0
 Export-Package:\
+	com.liferay.chat.constants,\
 	com.liferay.chat.exception,\
 	com.liferay.chat.model,\
 	com.liferay.chat.service,\
