+++ modules/apps/chat/chat-api/src/main/java/com/liferay/chat/constants/ChatPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.chat.web.constants;
+package com.liferay.chat.constants;
 
 /**
  * @author Peter Fellwock
