+++ modules/apps/chat/chat-service/src/main/java/com/liferay/chat/internal/poller/ChatPollerProcessor.java
@@ -15,13 +15,13 @@
 package com.liferay.chat.internal.poller;
 
 import com.liferay.chat.configuration.ChatGroupServiceConfiguration;
+import com.liferay.chat.constants.ChatPortletKeys;
 import com.liferay.chat.model.Entry;
 import com.liferay.chat.model.Status;
 import com.liferay.chat.service.EntryLocalServiceUtil;
 import com.liferay.chat.service.StatusLocalServiceUtil;
 import com.liferay.chat.util.BuddyFinderUtil;
 import com.liferay.chat.util.ChatConstants;
-import com.liferay.chat.web.constants.ChatPortletKeys;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.exception.NoSuchLayoutSetException;
 import com.liferay.portal.kernel.exception.NoSuchUserException;
